pyproject.toml
[project]
name = "tcb-text-to-graphql"
version = "0.1.0"
description = "Text-to-GraphQL AI Assistant PoC for Banking Data Lake"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.4.0",
    "langgraph-supervisor>=0.0.10",
    "chromadb>=0.6.0",
    "strawberry-graphql[fastapi]>=0.260.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "gradio>=5.0.0",
    "llama-index>=0.12.0",
    "llama-index-vector-stores-chroma>=0.4.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-llms-openai>=0.3.0",
    "llama-index-readers-file>=0.4.0",
    "llama-index-retrievers-bm25>=0.4.0",
    "tiktoken>=0.8.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "unidecode>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
